+++ NEWS.md
@@ -1,5 +1,8 @@
 # teal.slice 0.1.1.9015
 
+### New features
+
+* Added `set_filterable_varnames` method to `FilteredData` and `FilteredDataset` objects so that there is control over which variables can be filtered.
 * Removed `CDISCFilteredDataset` class and functionality moved to `CDISCFilteredData`.
 * Changed constructor of `FilteredData` to not require `TealData` object. See `help(init_filtered_data)` for more details.
 * The filtered data is now stored in `FilteredData` not `FilteredDataset`.
@@ -7,7 +10,11 @@
 * Updated `get_filter_state` to return a list of active filter states and an attribute with the character form of the filter states.
 * Updated `get_varlabels` method in the `FilterStates` classes, accepts a vector input.
 
-# Bug fixes
+### Breaking changes
+
+* Renamed internal S3 method `get_filterable_varnames` to `get_supported_filter_varnames`.
+
+### Bug fixes
 
 * Fixed a bug when the filter panel overview would not refresh if the panel was hidden during a transition between active modules.